requirements.txt
aiogram==3.15.0
aiohttp==3.11.11
openai==1.58.1
python-dotenv==1.0.1
